remove redundant wizard step
